Move the process step definition to helpers
